changed share number type to double

+++ classes/Position.h
@@ -6,7 +6,7 @@
 
 struct Position {
   Stock &stock;
-  int shares = 0;
+  double shares = 0;
 
   [[nodiscard]] double value() const;
 };
